Change fee percentage to float

+++ hestia/models.go
@@ -45,8 +45,8 @@ type Payment struct {
 }
 
 type Properties struct {
-	FeePercentage int  `firestore:"fee_percentage" json:"fee_percentage"`
-	Available     bool `firestore:"available" json:"available"`
+	FeePercentage float64 `firestore:"fee_percentage" json:"fee_percentage"`
+	Available     bool    `firestore:"available" json:"available"`
 }
 
 type Config struct {
